refactor: add dir structure to policy outputs folder
