--- src/pages/policy/output/DownloadableScreenshottable.jsx
@@ -1,13 +1,0 @@
-import Screenshottable from "../../../layout/Screenshottable";
-import React from "react";
-
-const DownloadableScreenshottable = React.forwardRef((props, ref) => {
-  return (
-    <div id="downloadable-content" ref={ref}>
-      <Screenshottable {...props} />
-    </div>
-  );
-});
-DownloadableScreenshottable.displayName = "DownloadableScreenshottable";
-
-export default DownloadableScreenshottable;
