+++ src/pages/policy/output/ImpactTypes.jsx
@@ -1,18 +1,18 @@
-import averageImpactByDecile from "./AverageImpactByDecile";
-import averageImpactByWealthDecile from "./AverageImpactByWealthDecile";
-import budgetaryImpact from "./BudgetaryImpact";
+import averageImpactByDecile from "./decile/AverageImpactByDecile";
+import averageImpactByWealthDecile from "./decile/AverageImpactByWealthDecile";
+import budgetaryImpact from "./budget/BudgetaryImpact";
 // import cliffImpact from "./CliffImpact";
-import deepPovertyImpact from "./DeepPovertyImpact";
-import deepPovertyImpactByGender from "./DeepPovertyImpactByGender";
-import detailedBudgetaryImpact from "./DetailedBudgetaryImpact";
+import deepPovertyImpact from "./poverty/DeepPovertyImpact";
+import deepPovertyImpactByGender from "./poverty/DeepPovertyImpactByGender";
+import detailedBudgetaryImpact from "./budget/DetailedBudgetaryImpact";
 import inequalityImpact from "./InequalityImpact";
-import intraDecileImpact from "./IntraDecileImpact";
-import intraWealthDecileImpact from "./IntraWealthDecileImpact";
-import povertyImpact from "./PovertyImpact";
-import povertyImpactByGender from "./PovertyImpactByGender";
-import povertyImpactByRace from "./PovertyImpactByRace";
-import relativeImpactByDecile from "./RelativeImpactByDecile";
-import relativeImpactByWealthDecile from "./RelativeImpactByWealthDecile";
+import intraDecileImpact from "./decile/IntraDecileImpact";
+import intraWealthDecileImpact from "./decile/IntraWealthDecileImpact";
+import povertyImpact from "./poverty/PovertyImpact";
+import povertyImpactByGender from "./poverty/PovertyImpactByGender";
+import povertyImpactByRace from "./poverty/PovertyImpactByRace";
+import relativeImpactByDecile from "./decile/RelativeImpactByDecile";
+import relativeImpactByWealthDecile from "./decile/RelativeImpactByWealthDecile";
 import laborSupplyImpact from "./LaborSupplyResponse";
 
 const map = {
