+++ src/pages/policy/output/decile/AverageImpactByWealthDecile.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import ImpactChart from "./ImpactChart";
-import { ImpactPlot, title } from "./RelativeImpactByDecile";
+import ImpactChart from "../ImpactChart";
+import { ImpactPlot, title } from "./AverageImpactByDecile";
 
 const description = (countryId) => (
   <p>
@@ -10,11 +10,11 @@ const description = (countryId) => (
   </p>
 );
 
-export default function relativeImpactByWealthDecile(props) {
+export default function averageImpactByWealthDecile(props) {
   const { impact, policyLabel, metadata, mobile, useHoverCard = false } = props;
-  const wealthDecileRelative = impact.wealth_decile.relative;
+  const wealthDecileAverage = impact.wealth_decile.average;
   const relativeChange =
-    -impact.budget.budgetary_impact / impact.budget.baseline_net_income;
+    -impact.budget.budgetary_impact / impact.budget.households;
   const chart = (
     <ImpactChart
       title={title(policyLabel, relativeChange, metadata)}
@@ -23,7 +23,7 @@ export default function relativeImpactByWealthDecile(props) {
       <ImpactPlot
         decileType={"wealth decile"}
         xaxisTitle={"Wealth decile"}
-        decileRelative={wealthDecileRelative}
+        decileAverage={wealthDecileAverage}
         metadata={metadata}
         mobile={mobile}
         useHoverCard={useHoverCard}
@@ -31,15 +31,9 @@ export default function relativeImpactByWealthDecile(props) {
     </ImpactChart>
   );
   const csv = () => {
-    const header = ["Wealth Decile", "Relative Change"];
-    const data = [
-      header,
-      ...Object.entries(wealthDecileRelative).map(
-        ([decile, relativeChange]) => {
-          return [decile, relativeChange];
-        },
-      ),
-    ];
+    const data = Object.entries(impact.wealth_decile.average).map(
+      ([key, value]) => [`Decile ${key}`, value],
+    );
     return data;
   };
   return { chart: chart, csv: csv };
